requirements.txt
tensorflow==2.12.*
tensorflow-hub==0.15.*
tensorflow-text==2.12.*
tf-models-official==2.12.*
matplotlib>=3.7
google-cloud-aiplatform>=1.38
google-cloud-storage>=2.10
